WIP: chore: macOS app bundling and packaging
Implement a `pkg-macos` target that ultimately results in `Status.dmg` being
written to `pkg/`.

Refactor a number of sections in the Makefile for consistency's sake, e.g. the
`appimage` target becomes `pkg-linux`.

This commit is WIP because it makes changes that result in the `build-linux`
target not working as is. The work here needs to be reconciled with #82 and
e.g. [nim-status-client/blob/windows-2/Makefile#L95-L103][build-linux-refactor].

In particular, a static build of DOtherSide is used and the Qt
libs ("frameworks" on macOS) are explicitly linked.

It's important that `PATH` and `QTDIR` environment variables are setup
correctly before running `make` (with or without an explicit target). On my
machine I do this when opening a terminal to use for building and packaging:

```sh
export PATH="${HOME}/Qt/5.14.2/clang_64/bin:${PATH}"
export QTDIR="${HOME}/Qt/5.14.2"
```

`pkg-macos` creates an `.app` bundle layout in `tmp/macos/dist` based partly on
information in a very helpful [StackOverflow answer][so-answer]. Note the part
of the answer (toward the end) that explains a problem with the working
directory and how to fix it. That's the reason for the `nim_status_client.sh`
script introduced in this commit (it gets copied into the bundle). It's also
the reason for having `Info.plist` copied into the bundle before `macdeployqt`
is run (see below) and then overwriting it with `Info.runner.plist` before
creating the `.dmg` file. The app icons file `status-icon.icns` was taken from
`deployment/macos/` in the [status-react][sr] repo.

The [`macdeployqt`][macdeployqt] tool is used to copy the needed portions of Qt
into the bundle; it automatically updates `rpath`, etc. so the
`nim_status_client` executable in the bundle can locate the libs within the
bundle.

The [create-dmg][cdmg] tool is used to package the bundle into a `.dmg`
file. There are two reasons for this:
1. It produces a nice looking icon for the `.dmg` that overlays the Status logo
on an external disk icon.
2. `Info.plist` needs to be overwritten after running `macdeployqt` (see
explanation above) but before creating the `.dmg` file. If we passed the `-dmg`
cli option to `macdeployqt` to have it generate the `.dmg` file then it
wouldn't be possible to overwrite `Info.plist`.

So there is a cosmetic reason and a practical reason for using another
tool. Probably the biggest downside is that `create-dmg` is implemented in
Node.js so it needs to be installed with `npm`; that's the reason this commit
introduces `package.json`, etc. Note that zero code from `node_modules` ends up
in the `.app` bundle or the `.dmg` file.

[build-linux-refactor]: https://github.com/status-im/nim-status-client/blob/windows-2/Makefile#L95-L103
[so-answer]: https://stackoverflow.com/a/3251285
[macdeployqt]: https://doc.qt.io/qt-5/macos-deployment.html
[sr]: https://github.com/status-im/status-react/tree/develop/deployment/macos
[cdmg]: https://github.com/sindresorhus/create-dmg