fix: push notification even when message comes from active channel
Prior to this commit, we would not push a notification when the message signal
has the same chat id as the currently active channel.

This caused some unexpected behaviour in which notifications wouldn't be shown
because the messages are received in the active channel, but the app may not be
"active" and in the background.

Another scenario where we need to ensure to push a notification is when a user
is using a different view than the chat (e.g. the wallet). Also here, a notification
should be pushed, even when the message channel is the same as the current active one.